Fix build with recent MinGW compilers by updating dlfcn-win32
Recent MinGW compilers fail to export the public symbols in dlfcn-win32.h. This
results in missing symbols while linking, thus failing the build. This is fixed
in the upstream dlfcn-win32 project at https://github.com/dlfcn-win32/dlfcn-win32.
Update our copy with the current upstream version fixes this so it's possible
to compile the addon with current MinGW compiler.