Bundle MinGW dependency DLLs with the addon
When building with the current MinGW compilers the addon DLL has runtime
dependency on several MinGW DLLs. These DLLs should be included in the addon
directory in order for the addon DLL to load. Currently only direct
dependencies of the addon DLL are included, no recursive dependency check for
the bundled DLLs is done. However, this appears to be ok for now.

Tested with official Kodi 15.2 build on Windows 10 x64 with MinGW 4.0 runtime
as installed by Cygwin and addon cross-compiled with the current MinGW
compilers in Cygwin (based on GCC 4.9.3) and Ubuntu wily (based on GCC 4.9.2).